+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Changes
 - Improved `alias.go` files (#103)
 - Added the support for posting empty-message posts with medias (#110)
+- Implemented the support for hashtags in posts (#96)
 
 ## Bug fixes
 - Fixed a bug inside the migration procedure of the `magpie` module (#106)